+++ src/lf_py_stack/orchestration/__init__.py
@@ -1,5 +1,7 @@
 from .cli import cli_app
-from .components import run_cli_command, truncate
+from .components import get_dbt_versions, log_dbt_versions, run_cli_command, truncate
+from .config import config
+from .duckdb import shrink_duckdb
 from .logger import get_logger
 from .mail import mail_app, send_mail
 from .steps import StepResult, log_step_nodes_table, log_step_results_table
@@ -8,10 +10,14 @@ __all__ = [
     "StepResult",
     "log_step_nodes_table",
     "log_step_results_table",
+    "log_dbt_versions",
+    "get_dbt_versions",
     "run_cli_command",
     "get_logger",
     "send_mail",
+    "shrink_duckdb",
     "truncate",
     "mail_app",
     "cli_app",
+    "config",
 ]